@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import "./index.css";
 import App from "./App";
 import { ChakraProvider } from "@chakra-ui/react";
@@ -9,7 +9,10 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 const queryClient = new QueryClient();
 
-ReactDOM.render(
+const container = document.getElementById("root")
+const root = createRoot(container)
+
+root.render(
   <React.StrictMode>
     <ChakraProvider theme={theme}>
       <QueryClientProvider client={queryClient}>
@@ -19,5 +22,4 @@ ReactDOM.render(
       </QueryClientProvider>
     </ChakraProvider>
   </React.StrictMode>,
-  document.getElementById("root")
 );
