@@ -1,5 +1,5 @@
 import axios from 'axios';
-const BASE_URL = 'https://basketful.onrender.com';
+const BASE_URL = 'http://localhost:8000';
 
 export default axios.create({
     baseURL: BASE_URL
