@@ -1,19 +1,32 @@
-import React, { useState } from "react";
-import { Flex, Heading, Button } from "@chakra-ui/react";
+import React, { useEffect, useState } from "react";
+import {
+  Flex,
+  Heading,
+  Button,
+  useBreakpointValue,
+  Icon,
+} from "@chakra-ui/react";
 import { Link } from "react-router-dom";
 import ProductItem from "./ProductItem";
 
-const Products = ({ title, items, ...props }) => {
+import { ReactComponent as LeftArrow } from "../../assets/arrow-dark-left.svg";
+import { ReactComponent as RightArrow } from "../../assets/arrow-dark-right.svg";
+
+const Products = ({ title, items, loading, maxItemsPerView, ...props }) => {
   const [slideIndex, setSlideIndex] = useState(0);
+  const itemsPerView = useBreakpointValue({
+    base: items.length,
+    lg: maxItemsPerView,
+  });
+
   return (
     <Flex
       flexDir="column"
       py="2em"
       // ml={{ sm: "158px", lg: "0" }}
       mx="auto"
-      px={{ base: "1rem", lg: 0}}
+      px={{ base: "5px", xl: 0 }}
       w="100%"
-      overflow="hidden"
       gap="1rem"
       {...props}
     >
@@ -27,21 +40,27 @@ const Products = ({ title, items, ...props }) => {
         </Heading>
         <Link to="">See All</Link>
       </Flex>
-      <Flex 
-      position="relative" 
-      justifyContent="space-between" 
-      gap="1rem"
-      overflowX="scroll"
-      className="categories-scorllable"
+      <Flex
+        position="relative"
+        justifyContent="space-between"
+        overflowX={{ base: "scroll", lg: "hidden" }}
+        className="categories-scrollable"
+        gap="10px"
       >
-        {items.map((item, index) => (
-          <ProductItem key={index} {...item} />
-        ))}
+        {loading
+          ? [...Array(6)].map((_, index) => (
+              <ProductItem isSkeleton={true} key={index} />
+            ))
+          : items
+              .slice(slideIndex, slideIndex + itemsPerView)
+              .map((item, index) => <ProductItem key={index} {...item} />)}
         <Button
+          display={items.length <= slideIndex + itemsPerView ? "none" : "unset"}
           position="absolute"
           top="50%"
           transform="translateY(-100%)"
-          right="-62px"
+          // right="-62px"
+          right="0"
           bg="none"
           _active={{ bg: "none" }}
           _hover={{ bg: "none" }}
@@ -49,29 +68,30 @@ const Products = ({ title, items, ...props }) => {
           p="16px"
           color={slideIndex === items.length ? "primary.500" : "dark.100"}
           fontWeight="bold"
-          fontSize="18px"
+          fontSize="32px"
           transition="0.6s ease"
           borderRadius="0 3px 3px 0"
           userSelect="none"
           onClick={() => {
-            if (items.length - (slideIndex + 4) < 4) {
-              setSlideIndex(items.length - (slideIndex + 4));
+            if (items.length > slideIndex + itemsPerView) {
+              setSlideIndex(items.length - (slideIndex + itemsPerView));
             }
           }}
         >
-          &#10095;
+          <Icon as={RightArrow} h={10} w="auto" />
         </Button>
         <Button
-          m={5}
+          display={slideIndex === 0 ? "none" : "unset"}
           position="absolute"
           top="50%"
           transform="translateY(-100%)"
-          left="-84px"
+          // left="-84px"
+          left="0"
           width="auto"
           p="16px"
           color={slideIndex > 3 ? "primary.500" : "dark.100"}
           fontWeight="bold"
-          fontSize="18px"
+          fontSize="32px"
           transition="0.6s ease"
           borderRadius="3px 0 0 3px"
           userSelect="none"
@@ -79,13 +99,12 @@ const Products = ({ title, items, ...props }) => {
           _active={{ bg: "none" }}
           _hover={{ bg: "none" }}
           onClick={() => {
-            console.log(slideIndex);
-            if (slideIndex > 3 && slideIndex - 4 <= 0) {
-              setSlideIndex(slideIndex - (slideIndex - 4));
-            }
+            setSlideIndex(
+              slideIndex - itemsPerView > 0 ? slideIndex - itemsPerView : 0
+            );
           }}
         >
-          &#10094;
+          <Icon as={LeftArrow} h={10} w="auto" />
         </Button>
       </Flex>
     </Flex>
